loading .env variables in scripts

+++ PRA04_ANSWER/copy-resources.sh
@@ -1,8 +1,17 @@
 #!/bin/bash
 
+ENV_VARS_TO_LOAD=$(bash util/clean-env-file.sh .env)
+if [ $? -ne 0 ]; then
+    exit 1
+fi
+if [ -z "$ENV_VARS_TO_LOAD" ]; then
+    echo ".env file has no var defined!"
+    exit 1
+fi
 while IFS== read -r key value; do
     printf -v "$key" %s "$value" && export "$key"
-done <.env
+done <<< "${ENV_VARS_TO_LOAD}"
+
 
 rm -rf "${SPRING_DIR}"
 mkdir -p "${SPRING_DIR}"
